+++ ntex-net/CHANGES.md
@@ -8,6 +8,8 @@
 
 * Update IoContext::update_read_status() api usage
 
+* Update IoContext::update_write_status() api usage
+
 * Produce io::ErrorKind::WriteZero for backend impl
 
 ## [4.0.1] - 2026-09-18